Move toggle add order test to OpenOrder spec
